Filenet socket rework.
There was a rather large problem where the socket for a fn
connection was stored in the fnetnode itself, and filenet.c had
opinions on clearing its data internally in killfnetnode, which
led to a number of consistency problems -- the code was ugly,
and the daemon would occasionally segfault if killfnetnode had
cleared sk->data prematurely for some reason.

Now, each fnet module takes care of the socket itself, so that
it is modified internally only. The code is much cleaner that
way.